fix(session): free client-side sessions if no longer needed
Before this commit, client-side sessions were kept in a Vec inside of
the CoapContext for the entire lifetime of the context instead of
freeing them as soon as they were dropped.

This commit fixes the issue by fixing the Drop implementation of
CoapClientSessionInner, which caused a SIGABRT if it was called outside
of the CoapContext drop handler.

Additionally, client-side sessions are no longer stored in the
CoapContext, as they were not used there except for in the Drop
implementation, which is no longer necessary due to the aforementioned
fix.