fix(server): Don't send a copy of the PDU instead of applying the response to the original response PDU
This will require application-side modifications to server-side request handlers.
Namely, the resource handler must no longer call coap_send for the response PDU.

This change was made to conform to libcoap's API contract and to enable CoAP Observe support.

Additionally, a small workaround for proper CoAP Observe support was added that prevents duplicate
addition of the Observe option in server-side resource handlers.

This workaround should be removed once #21 has been tackled and the API for message handling was reworked.